added notifications
with the integration of a webserver, notifications have been able to be
added. webserver has its own github repo. (can be found under my
account)

the issues:
- webserver is residing on localhost, emulator cannot properly discover
the localhost location as it creates its own 'localhost' secluding
itself from the server. thus the server's ip (run locally on my
computer), my computer's ip, is being hardcoded in

- notifications have to be manually pushed.. this was unfortunately
expected due to data being driven by the app and not a server.. (looking
back on it .. a poor choice)

- notifications don't properly link anywhere.. just kind of popup, that
really just an undersight on my part.

-